requirements.txt
llama-cpp-python=0.2.78
langchain=0.2.5

